Implimenting body part disablement
started putting in calls to be able to detect health dropping bellow
body part max, then calling a 'disable' of all the squares associated
with it. Only left arm works consistently. Call for checking to disable
body parts is in gamecontroller update(). Might not be the best place.

Need to revise and edit calls and methods, getting too disorganized

+++ Basic Cards/Assets/Scripts/CardScript.cs
@@ -7,7 +7,7 @@ public class CardScript : MonoBehaviour {
 	int cardSpriteNum;
 	string nameOfCard; 
 	int rankOfCard; 
-	int attackDamageOfCard; 
+	float attackDamageOfCard; 
 	string typeOfAttack;
 
 	private DeckScript deckScript;
@@ -62,7 +62,7 @@ public class CardScript : MonoBehaviour {
 	public int CardNumber{
 		get{return cardSpriteNum;}
 	}
-	public int AttackValue{
+	public float AttackValue{
 		get{return attackDamageOfCard;}
 	}
 	public string TypeOfAttack{
